Added check in to exists template and export_path

+++ lib/planetarium/command.rb
@@ -22,28 +22,34 @@ module Planetarium
         template = options[:template]
       end
       
-      feeds = Feedzirra::Feed.fetch_and_parse(config.urls).each_value
+      if !File.exists? export_path
+        puts "ERROR: export_path (#{export_path}) don't exists"
+      elsif !File.exists? "#{TEMPLATE_PATH}/#{template}"
+        puts "ERROR: template (#{template}) don't exists"
+      else
+        feeds = Feedzirra::Feed.fetch_and_parse(config.urls).each_value
 
-      # loading plugins
-      # TODO: - improve load
-      #       - loader plugin with marked enable
-      plugins = {}
-      Dir.glob("#{File.dirname(__FILE__)}/plugins/*.rb").each { |file| require file }
-      Planetarium::Plugin::Base.repository.each do |r|
-        p = r.new
-        plugins[p.plugin_name] = p
-      end
+        # loading plugins
+        # TODO: - improve load
+        #       - loader plugin with marked enable
+        plugins = {}
+        Dir.glob("#{File.dirname(__FILE__)}/plugins/*.rb").each { |file| require file }
+        Planetarium::Plugin::Base.repository.each do |r|
+          p = r.new
+          plugins[p.plugin_name] = p
+        end
 
-      # export html
-      File.open("#{export_path}/index.html", "w") do |out|
-        t = ERB.new(File.read("#{TEMPLATE_PATH}/#{template}/index.html.erb"), 0, "%<>")
-        out.puts t.result(binding)
-      end
+        # export html
+        File.open("#{export_path}/index.html", "w") do |out|
+          t = ERB.new(File.read("#{TEMPLATE_PATH}/#{template}/index.html.erb"), 0, "%<>")
+          out.puts t.result(binding)
+        end
 
-      # export atom
-      File.open("#{export_path}/atom.xml", "w") do |out|
-        t = ERB.new(File.read("#{TEMPLATE_PATH}/atom.xml.erb"), 0, "%<>")
-        out.puts t.result(binding)
+        # export atom
+        File.open("#{export_path}/atom.xml", "w") do |out|
+          t = ERB.new(File.read("#{TEMPLATE_PATH}/atom.xml.erb"), 0, "%<>")
+          out.puts t.result(binding)
+        end
       end
     end
